# packages/oic-client/src/generated/oic-openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Oi Intelligence Core native and compatibility API",
    "title": "OIC API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/admin/applications": {
      "post": {
        "operationId": "FoundationController_createApplication_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/by-key/{key}": {
      "get": {
        "operationId": "FoundationController_getApplicationByKey_v1",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}": {
      "get": {
        "operationId": "FoundationController_getApplication_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/credentials/{credentialId}": {
      "delete": {
        "operationId": "FoundationController_revokeCredential_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "credentialId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/external-references/{referenceId}": {
      "delete": {
        "operationId": "FoundationController_revokeExternalReference_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "referenceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      },
      "patch": {
        "operationId": "FoundationController_remapExternalReference_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "referenceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
...
    "/api/v1/admin/applications/{applicationId}/service-principals": {
      "post": {
        "operationId": "FoundationController_createPrincipal_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/service-principals/{principalId}": {
      "get": {
        "operationId": "FoundationController_getPrincipal_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "principalId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/service-principals/{principalId}/credentials": {
      "post": {
        "operationId": "FoundationController_issueCredential_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "principalId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/service-principals/{principalId}/scopes": {
      "delete": {
        "operationId": "FoundationController_revokeScope_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "principalId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      },
      "post": {
        "operationId": "FoundationController_grantScope_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "principalId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/service-principals/{principalId}/tenants/{tenantId}/grant": {
      "delete": {
        "operationId": "FoundationController_revokeTenant_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "principalId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      },
      "post": {
        "operationId": "FoundationController_grantTenant_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "principalId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/tenants": {
      "post": {
        "operationId": "FoundationController_createTenant_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/tenants/by-external-reference/{sourceType}/{externalId}": {
      "get": {
        "operationId": "FoundationController_getTenantByExternalReference_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sourceType",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "externalId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/tenants/{tenantId}/external-references": {
      "post": {
        "operationId": "FoundationController_createExternalReference_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/applications/{applicationId}/{kind}/{id}/status": {
      "post": {
        "operationId": "FoundationController_changeLifecycle_v1",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "operationId": "FoundationController_readAudit_v1",
        "parameters": [
          {
            "in": "query",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
...
    "/api/v1/admin/tenants/{tenantId}": {
      "get": {
        "operationId": "FoundationController_getTenant_v1",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "OIC Foundation Administration"
        ]
      }
    },
...
  },
  "servers": [],
...
}